ci: more automated validation (#204)
* ci: run pod lib lint on release branches

* ci: check Carthage version link

* ci: accept Carthage JSON specification file as version-check script argument

* ci: cleanup and cosolidate build-and-test action

* chore: remove xcframework zip

* fix: set example test target deployment version 14.0, otherwise Mac Catalyst errors out

* ci: add barebones SPM example test

+++ scripts/run-tests-cocoapods-monterey.sh
@@ -1,17 +1,33 @@
 #!/bin/bash
 set -euo pipefail
 
-brew install xcbeautify
+readonly XCODE=$(xcodebuild -version | grep Xcode | cut -d " " -f2)
 
-echo "Running unit tests on Xcode version: $(xcode-select -p)"
+if ! command -v xcbeautify &> /dev/null
+then
+    echo -e "\033[1;31m ERROR: xcbeautify could not be found please install it... \033[0m"
+    exit 1
+fi
 
-# Delete the old stuff
+echo "▸ Current Xcode: $(xcode-select -p)"
+
+echo "▸ Using Xcode Version: ${XCODE}"
+
+echo "▸ Available Xcode SDKs"
+xcodebuild -showsdks
+
+echo "▸ Removing XCFramework folder"
 rm -Rf XCFramework
-# reset simulators
+
+echo "▸ Shutdown all simulators"
 xcrun -v simctl shutdown all
+
+echo "▸ Erase all simulators"
 xcrun -v simctl erase all
 
+echo "▸ Unzipping XCFramework"
 unzip MUXSDKStats.xcframework.zip
+
 cd apps/DemoApp
 
 echo "▸ Reset Local Cocoapod Cache"
