+++ scripts/run-tests-cocoapods-ventura.sh
@@ -1,17 +1,31 @@
 #!/bin/bash
 set -euo pipefail
 
-brew install xcbeautify
+readonly XCODE=$(xcodebuild -version | grep Xcode | cut -d " " -f2)
 
-# Delete the old stuff
+if ! command -v xcbeautify &> /dev/null
+then
+    echo -e "\033[1;31m ERROR: xcbeautify could not be found please install it... \033[0m"
+    exit 1
+fi
+
+echo "▸ Current Xcode: $(xcode-select -p)"
+
+echo "▸ Using Xcode Version: ${XCODE}"
+
+echo "▸ Available Xcode SDKs"
+xcodebuild -showsdks
+
+echo "▸ Removing XCFramework folder"
 rm -Rf XCFramework
-# reset simulators
+
+echo "▸ Shutdown all simulators"
 xcrun -v simctl shutdown all
+
+echo "▸ Erase all simulators"
 xcrun -v simctl erase all
 
-echo "test xcode version"
-xcodebuild -version
-
+echo "▸ Unzipping XCFramework"
 unzip MUXSDKStats.xcframework.zip
 
 cd apps/DemoApp
